pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "claude-runner"
version = "0.1.0"
description = "Token-budgeted to-do list runner for the Claude Agent SDK."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "Bill Denney", email = "wdenney@humanpredictions.com" }]
keywords = ["claude", "anthropic", "agent", "sdk", "scheduler", "queue"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "rich>=13.7",
    "platformdirs>=4.2",
]

[project.optional-dependencies]
api = ["anthropic>=0.40"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "time-machine>=2.14",
    "types-PyYAML",
]

[project.scripts]
claude-runner = "claude_runner.cli:main"
claude_runner = "claude_runner.cli:main"

[project.urls]
Homepage = "https://github.com/wdenney/claude_runner"
Issues = "https://github.com/wdenney/claude_runner/issues"
Changelog = "https://github.com/wdenney/claude_runner/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_runner"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pycodestyle / pyflakes
    "I",             # isort
    "B",             # bugbear
    "UP",            # pyupgrade
    "N",             # pep8-naming
    "SIM",           # simplify
    "RUF",           # ruff-specific
]
ignore = [
    "E501",   # line length is already flagged by formatter
    "UP042",  # str-enum is the pattern we actually use for stable wire values
    "B008",
    "SIM108",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["N802", "B007", "SIM117"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
show_error_codes = true
files = ["src/claude_runner"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=claude_runner",
    "--cov-report=term-missing",
]

[tool.coverage.run]
branch = true
source = ["claude_runner"]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "\\.\\.\\.",
]
